pyproject.toml
[project]
name = "maya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=0.3.33",
    "langchain-postgres>=0.0.17",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-supervisor>=0.0.29",
    "langmem>=0.0.29",
    "numpy>=2.3.3",
    "psycopg>=3.2.10",
    "pypdf>=6.7.5",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.2.3",
    "setuptools<81",
    "trustcall>=0.0.39",
]

